Update patch set 1
Patch Set 1:

Is there a way to make this configurable? It also appears to affect the screen when plugging in as well. It's inconvenient to have the screen come on when plugging in, especially during quiet hours.

Patch-set: 1
Reviewer: Captain Throwback <6558@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0