Update patch set 1
Patch Set 1: Code-Review-1

Actually the "standard" behavior (judging by nexus device behavior) should be to turn the screen on when unplugging. I'm not going to revert this, rather I would prefer to see the default overlay option turn screen on when unplugging instead of leaving it up to devices to do it.

Patch-set: 1
Label: Code-Review=-1