Update patch set 1
Patch Set 1: Code-Review-1

I prefer it on by default. Adding a settings to toggle it sounds like a good idea on the outside, but we will *not* be doing a DeviceParts style setting for it. This will have to be added into Settings for all devices (or Developer Options). I personally will never toggle it off, so having the toggle in is not a priority for me. Leaving this turned on, however, is. Even Other Nexus devices and devices with LEDs still have this enabled....so we should as well. (Check out sgs3, sgs4, moto devices, etc). Good idea to bring this up, but I do not want this reverted personally.

Patch-set: 1
Label: Code-Review=-1