Update patch set 1
Patch Set 1:

This is the initial changed based on what was done to m7. The remaining pieces broken on the HTC s4's are camera and bluetooth. Both of them are being denied access to run. Once this selinux policy is hashed out, both camera and bluetooth should begin working again.

Patch-set: 1