Update patch set 2
Patch Set 2:

@Captain:
You need to check what is the highest value for the currents file.
For m7, defualt is 20, but there is absolutly no change from 10-20, so I took 10 as max. That is why I devided by 25 because slider in Settings goes from 0-255.

So maybe your max is 20 so you need to divide by 10 or 12 to get it working throughout the scale of the slider

Patch-set: 2