Update patch set 3
Patch Set 3: Code-Review+1

I tested this on fireball and verified the cpu governor changes during and after backup. The performance gains are minimal, but we should be doing users a favor by turning off one of the CPUs when it is unneeded. Menu navigation with swipe is still responsive with 1CPU in ondemand.

Patch-set: 3
Label: Code-Review=+1