Update patch set 1
Patch Set 1: Code-Review+1

Unless there are objections, I will merge this tonight. I haven't really done much with camera in the past, so this is my only hesitation in merging outright.

As a side note, it looks like we still need to look at touch-to-capture while video is recording. I get garbage pictures in that scenario.

Patch-set: 1