Update patch set 3
Patch Set 3: Code-Review+1

Re-uploaded to resolve merge conflict with the performance profiles overlay.

I'm not particularly thrilled about this for s4, but I admit even fireball handles adding a new user reasonably well, though it is slow. Storage is an issue, but I'm not going to block based on this. How users use their limited storage is up to them.

Patch-set: 3