Update patch set 1
Patch Set 1:

Sadly, it seems this is necessary on fireball again. I believe one of noobnl's dctracker commits actually prevented this issue in the past, but there are fewer of those commits in opt/telephony now.

Will hold off for a little while to see how things in opt/telephony shake out.

Patch-set: 1