Update patch set 1
Patch Set 1:

1) Cherry-picked http://review.cyanogenmod.org/92907 to cm-12.1

2) Modified commit to move "inline const LocPosMode& getPositionMode() const {return fixCriteria;}" below the ordered virtual enums, rather than delete it.

3) Experience: not 100% trouble free, but GPS is working on fireball:
GPS failed to see any satelites for a solid minute, so I cleared AGPS and then updated AGPS and *bam*, satellites.

I thought I would be able to scan the log for mentions of setPrivacy, but apparently these LOC_LOGWs are not logged anywhere?

Not sure whether this is going to be worth it in the end, but at least progress.

Patch-set: 1