Update patch set 1
Abandoned

ROW is up to date in this device's kernel, and the community still appears to be shaking-down the alternatives to ROW and CFQ. I'll let this one play out a bit long. In any case, we can set this via a system prop rather than in init.

Patch-set: 1
Status: abandoned