Update patch set 1
Patch Set 1:

As far as RIL_REQUESTs, RIL_NO_CELL_INFO_LIST is the only one we handle in libril. I'm actually interested in moving that change out of libril since it messes with indices. We could move the unsol remappings here as well. Ultimately, it would just make for fewer changes to keep track of when diff'ing libril, whereas java just gets an overlay.

Patch-set: 1