Update patch set 2
Patch Set 2:

(on the road) So, this commit shouldn't be needed, should it? As long as the radio capabilities request isn't rejected anymore (resolved by commit just ahead of this one), does frameworks handle creating a static list based on  config_radio_access_family? If this is the case, it can be reverted on m7-common.

Patch-set: 2