Update patch set 2
Patch Set 2:

There's a case where the RIL doesn't understand the message, but doesn't respond with an unknown message exception. This happened on some other devices where the OEM had their own message ID with the same ID as the request radio access family message. I'm not sure that's the case on HTC though.

Patch-set: 2