Update patch set 1
Patch Set 1: -Code-Review

The current version of Snap is crashing when switching from back to front cam, but I do not think it is related to this commit. I still need to test a bit more, but right now I'm thinking all legacy devices should go ahead and set these in their wrapper or hal.

Patch-set: 1