Update patch set 2
Patch Set 2:

My best guess about this denial is that it's due to the setprop in init.qcom.bt.sh. I wonder whether it's worth plucking what we need from that script, putting it in init.qcom.rc, and deleting the script. It's not like we're going to see any updates from qcom for bt on msm8960.

[ 1875.954738] type=1400 audit(1484847961.346:87): avc: denied { write } for pid=5349 comm="setprop" name="property_service" dev="tmpfs" ino=6797 scontext=u:r:toolbox:s0 tcontext=u:object_r:property_socket:s0 tclass=sock_file permissive=0

Patch-set: 2